Switch the Octokit dependency from `octokit` to `@octokit/rest`
We use Octokit to interact with the GitHub API. This switches
from the `octokit` package to the more specific `@octokit/rest`,
which avoids a dependency on `@octokit/webhooks`.

Avoiding this dependency is good because it in turns depends on
`clean-stack`, which doesn't play nicely in all JS runtimes - see
https://github.com/sindresorhus/clean-stack/issues/28.

This cleans up at least the immediate cause of #25.